# PERF: select_dtypes