```
In [1]: from asv_bench.benchmarks.dtypes import *

In [2]: cls = SelectDtypes

In [3]: self = cls()

In [4]: dtype = "Int32"

In [5]: self.setup(dtype)

In [6]: %timeit self.time_select_dtype_string_exclude(dtype)
1.79 ms ± 50.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)   # <-- master
47.7 µs ± 461 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <-- PR
```