@jreback 2021-07-19
nice, can you rebase (I think the typing issue is resolved on master)